requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
colorlog==6.6.0
frozenlist==1.3.0
hikari==2.0.0.dev106
hikari-lightbulb==2.2.0
idna==3.3
multidict==6.0.2
py-cord==2.0.0b4
pycodestyle==2.8.0
pycparser==2.21
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
toml==0.10.2
urllib3==1.26.8